Rename keras.py to kkeras to prevent import confusion
Since the keras package and the keras file had the same name, importing
keras was a little bit obscure: did we mean 'import the keras package' or
'import the keras.py file'?
